Aduino Oregon Scientific Weather Sensors Interceptor
Updated the weather station to include monitoring an
Oregon Scientific UV Detector UVN-800.
As well the promised RGB status indicator with a different colour
for each sensor logged, and white to signal data sent.
The system also flags when a sensor is missed for 20 minutes
in a row, and reset is at least 100 subsequent reads occur.
This allows monitoring of battery status and such as things
as antenna changes, or signal obstruction or RFI.